User Auth and Devise Layout
Now that a user is able to be signed in I wanted to ensure that the
internal application checks their authentication. I added the provided
devise method to the authenticated controller.

The Login page was then on the wrong layout so I used the devise custom
base class to ensure that we use the right layout for the devise views.

https://github.com/plataformatec/devise/wiki/How-To:-Create-custom-layouts