Fix Preferred Hours DataTypes
Time sort of sucked. It contained date info which I did not want. Thus I
am moving this to an integer column. This will allow it to be the start
time in minutes vs the end time in minutes.

Converting from Time to Integer = Impossible, so I had to drop and
recreated the columns.
http://stackoverflow.com/questions/5522520/rails-postgres-migration-to-change-colomn-gives-error-cannot-be-cast-to-type