Begin Adding Scheduling Rules
Companies have positions and rules around when those positions need
filled. The position model is made to introduce the rank of the
employee.

A scheduling rule is a basic rule to start to adjust the scheduling
algorithm. This is the simplest implementation but will be added to over
time. Right now the rule takes in the position, the num of employees and
the shift time (open, close, or all day).

Future customizations include adding a custom time (for places where the
hours matter) and adding in a busy / average / light workload.

The employer could then mark certain days / hours as busy and the
algorithm could adjust to that.