Use Parameter On Start of Week
Start of week and end of week take a symbol for the day that the week
starts on. Ruby defaults to monday because made in japan which means i
was subtracting a day. This caused a weird bug for months that end on a
sunday because it cut off the last day.

https://apidock.com/rails/v3.2.13/Date/end_of_week