Prevent Days Outside the month from highlighting
The calendar only checked to see if the day was the same in order to
highlight it. However since towards the end of the month there could be
an identical date in another month displayed it ended up highlighting
the grayed out date in the wrong month. IE March 29th and April 29th
were both displayed and highlighted.