Add Location and User Id to Shifts
In the UI things were getting complicated getting users and locations
through the user_locations table. While the data is normalized nicely it
starts to break down when we add and remove users from locations. This
creates more complicated shift handling and such. To make things simpler
this creates direct references to both location and user.

Note: The migration is not backwards compatible if there are shifts in
the table. You will need to delete all your current shifts first.