Build Locations Instead of Create
In order to ensure that the company settings are pulled in, we will
build instead of create. This ensures that when they do toggle off the
'use_company_settings' checkbox that the company settings of today are
the starting point.

Add in some attributes to the params whitelist as well as ensure that if
the use company settings box is checked to start the form is disabled.