Better Navigation To Get to Your Shifts Page
This is a pretty important page so I want to add to the top and left
menu. To save some space on the top I am removing that from the quick
list at the top and leaving it just on the side. I may add it to the
dropdown menu from the user at some point.