Track State Of Shifts
We need to start to track if a shift should not be shown typically. This
could happen for a variety of reasons like getting traded or cancelled
or taken over. We still want to retain the shift data for the original
user but will make a new shift for the new user.