First Pass at Taking a Shift Over
Lots going on here.

Availability Checker - Ensure hours are available to trade. If a user
goes over the allotted hours we are going to need some sort of special
handling with manager approval. For now, we will not let them do it.

State tracking on trades - better tracking of what state trades and
shifts are in.

Shift Taker Model - Will handle taking a shift, a bit different than the
shift trader which trades shifts.