Soft Delete Users
When deleting a user we do not want to get rid of them. That will mess
up all sorts of stuff down the road. We want to mark them as soft
deleted. However if they come back the email needs to be there for them
to create a new email.

In the employee remover we add a unique string to the front of the email
in a way we can get the original email back as well as updated the
deleted_at column with the timestamp.