Positions Step
A large refactor on some onboarding flow layouts.

Extracted the frame and added it to the layout so it doesn't have to be
duplicated everywhere.

A user goes from company to positions instead of directly to locations.
Here they can delete and add positions. I needed to add a soft delete to
positions so that we didn't have any broken has through links.

In addition, position is where a location admin or company admin will
live outside the original user.