Onboarding Finish And Blank Calendar
If a user has no location, the app should let them know. This makes the
calendars show page support a no location. When this happens it will
tell them to have their manager add them to a location.

This makes the last step of onboarding connect but at this time does not
make the schedules. It is going to be a bit larger of a refactor to make
that area work well so I skipped it for now but will revisit.