Credit Card Addition Error Validations
This... was.. a... pain..

Since stripe does not allow us to send to our servers this requires
front end validation. However then simple form gets all funky and
translations are hard to do. Thus I passed a block to simple form so I
could pre render the localized errors. The javascript then checks the
values and displays the errors.

Much more to do in regards to CC error handling but this is at least a
start.