Onboarding Contact Form and Email
While we are gaining our initial traction we will do all the work for
them. This introduces a form where they can enter their details and we
will reach out and get in touch with them. This sends an email to
support letting us know that there is a customer interested in trying
our product and getting set up.