Add A User Context that takes in location
This is going to be a magical thing about this app that will probably
break often and be frustrating to remember. However we need to auth
policies based on a location as well as a user. This is because a user
could admin one location and not another. If before calling authorize a
location is set via instance var, it will auth for that user and that location.