Add InProgressShift, Update Scheduler
I'm on the fence about in progress shifts. I could see adding a state to
the shifts table to account for unpublished. However I feel there will
be many bugs down the line if we go that route. Keeping the in progress
schedules seperate allows for a clearer seperation initially and
hopefully less bugs down the line.

Updated the scheduler to accept this as its input. The rest of the
options are driven off the scheduling_period.