Primary Position
As an employee you are typically given one position. You can fill in as
others. Thus the join table of positions will be the places you can fill
in as, and the primary_position_id is a direct reference to the
position.

This route makes it easier to build form, have less validations etc.
Also it will help to avoid n+1 queries in the scheduling engine as we
can load the companies position into a hash and access directly each
time instead of having to do a query in a loop.