Add Job to Update Billing
This is a nightly job that will update the subscription for all the
companies. I like the fact that its nightly so we don't need to call it
every time an employee gets added or removed. If we forget it will just
get swept up that evening.

In the future we could add some caching to figure out if we need to make
the call or not to stripe but when we do that we have way too many
customers =).