Clean Up Schedule Hours
The full crud was built for this which is fine but probably too
complicated for us right now. We are going to live under the assumption
that a business has an open and a close time once per day. These are
auto built now so we just need to display and allow them to edit on a
per day basis.