Onboarding Flow Updates
Locations is going to become the last step that sort of repeats in a
circle. Thus I am using continue to redirect back to locations and then
they can continue on. Added some flash messaging that a user was created
and added a back button so they can view all employees. Updated some
copy and added positions into the onboarding flow.