Invitation URLs
In order to help get people in the platform this creates an easy way
for managers to not have to worry about collecting information. They can
post this link or send it around and. Their employees can then simply
sign up with the link and get added to the location.