Remove Scheduling Periods, Add Edited State
It's been decided to move away from scheduling periods and keep things
strictly based on timeframes. This removes the lookup of the scheduling
periods.

It also introduces the ability of an edited shift vs a published shift.
Admins for locations will have access to the in_progress_shifts table
which could contain edits. The edited field is a way to track this and
display a different shade / outline if so.

When they click publish it will publish that set of shifts and it will
become unedited. Appropriate people will be notified and all that stuff.