Set Edited True when InProgress Shift Is Edited
When editing the inprogress shift we should set edited to true. This
makes the visual indicator show that the shift is changed. In the future
we can add more complex logic around this flag setting to ensure its
actually changed and hasn't been just saved but this works for a first
pass.