Begin Populating Repeating Shifts
Anytime we look up a shift via date ranges we need to check those dates
to see if there are any repeating shifts to be scheduled at that point.
This adds the structure required to begin populating this stuff as well
as additional namespacing organization around classes that deal with
shifts and time.