Edit Repeating Shifts
Allow repeating shifts to get properly updated. This is where things get
confusing but I think this is the top of the hill so to speak. When a
repeating shfit gets updated we need to do different things based upon
what the user wants. If they update all of them we need to continue to
populate in_progress_shifts as the new values but populate shifts as the
old values. To do this I added preview columns that we pull all in
progress shifts from.