Start Overtime Calculations
This allows us to calculate overtime and pay on a per shift level.
Overtime happens in states that have overtime after 8 or 12 hours or on
the 6th or 7th day etc. We can add custom processors in the lookup
class. Adds some columns required to make this all work properly.