Simplify Creating Offers
Offers were a bit complicated with trading and I want to simplify this a
bit. If someone doesn't offer a shift then it will just takeover the
shift vs offering a shift will ensure that a trade is made. Both should
be subject to approval if require from the settings.