Set Initial Users Position
The initial user needs to get their primary position set. Thus for the
first position they enter we prompt them for their position and after
that ask for other positions in their company. If the primary position
is blank when creating a position we set it to that to avoid running
into an issue where the user that signs up cannot be scheduled.