TimePickerRange Updates To Onboarding
I found a few bugs with the timepicker range input. When deselecting it
reverted to the numerical value since we did not set it via the
typeahead method. Also when entering gibberish it would allow that
input even though we only want to allow the selections in the dropdown.
It took a custom funciton to do that and reset the index if the val was
set to something weird.

Closes #117