Bugfix: Back Button From Onboarding -> Dashboard
The back button breaks if you push it after you update your schedule
setting since it uses a render in the create/update method. This adds a
show action which simply redirects the person back to the last step in
the onboarding for them to fill out again if they want.