Finish Up Country and State Select
Finish up the country and state select implementation. A small JS
helper was added to tie the two inputs together on forms where they are
both present. Stayed with the country select gem because the list was
generated however we do need to narrow that down a bit. Also added
states for the other countries that support states.