Redraw Ability + Better In Progress Alert
When rendering the whole calendar over again it had an unfortunate
effect of removing your scroll location. This adds a function to more
smartly redraw the calendar by taking the full html and slicing out the
major sections and reappending them. This ensures that since the
original container divs remain on screen their scroll positions also
persist.

Also cleans up the sidebar and allows a smarter in progress alert
notification to display.