Location Headers & Select Disabled Fix
Add the location to the locations page so you know that you are editing
a location.

Also holy moly safari you make things difficult. I think.... this
solution will work for all the input forms but man was it annoying.
Each select needs a wrapper div with an ::after because of course you
cant add an ::after to a select option. This allows us to float the
dropdown over the select menu.