Test Theory
The initial rates to the funnel haven't been great. One theory is that
asking for an email makes them less likely to go to that page and thus
less likely to signup. This switches to a button for now to see if
rates improve over the next day or so.