Set Up Reporting Intro Page
Sometimes a user may not have a location. This sets up a landing page
when a person hits that scenario that allows them to select the location
and proceed onto the rest of the reporting. Stubbed a route for the
time sheet page. This is going to be largely a rework of the current
stats page now that some more advanced calculation classes are in place
that do a similar thing.