Test Out Funnel Conversion Theory
Right now people are hitting the signup page and bouncing. My guess is
people are hitting this page with not enough context from the hero
section of the home page.

I am trying out redirecting people to the pricing page where it goes
into the features and such in more detail. Hopefully they see that get
an idea and then continue on with the flow.