Rebuild User Profile With New Form Styling
This was supposed to be simple but introduces a whole bunch of new form
styles and elements including tag select and horizontal forms. In
addition there is an editing frame that can show and hide the
show/update modes in one form. This created the need for a whole bunch
of remote actions and slowed down development considerably.