Search Updates & Blank Option For Reporting
Figure out elasticsearch settings. The key is to use a ngram tokenizer
with no fuzziness. This makes an exact match required to an ngram
instead of substituting parts of the search term.

Moved public companies to the root so it can be created without a
company. This allows us to offer the option to make a report even if
the company is not there.